refactor: déplace les interfaces pures dans external
supprime searchable/ et selector/ qui ne sont plus utilisés
