+++ pfmg/lexique/paradigm/Paradigm.py
@@ -11,12 +11,12 @@ from pathlib import Path
 from typing import ClassVar
 
 from pfmg.external.reader.ABCReader import ABCReader
+from pfmg.external.realizable.ABCRealizable import ABCRealizable
 from pfmg.lexique.block.BlockEntry import BlockEntry
 from pfmg.lexique.forme.Forme import Forme
 from pfmg.lexique.forme.FormeEntry import FormeEntry
 from pfmg.lexique.lexeme.Lexeme import Lexeme
 from pfmg.lexique.morpheme.Morphemes import Morphemes
-from pfmg.lexique.realizable.ABCRealizable import ABCRealizable
 from pfmg.lexique.sigma import new_gloses
 from pfmg.lexique.sigma.Sigma import Sigma
 from pfmg.lexique.sigma.StraightPos2Sigmas import StraightPos2Sigmas
